тут выяснилось что отрицательный видс нужно обрабатывать... (683/694)

+++ lets_check_what_you_pars.c
@@ -6,7 +6,7 @@
 /*   By: cveeta <cveeta@student.21-school.ru>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/25 12:16:42 by cveeta            #+#    #+#             */
-/*   Updated: 2020/12/02 20:28:52 by cveeta           ###   ########.fr       */
+/*   Updated: 2020/12/02 20:32:31 by cveeta           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,5 +31,5 @@ void 	lets_check_what_you_fuck_SpArSiL(s_s_f *struct_, va_list *ap)
 	if (struct_->type == 'X')
 		write_type_x_X(va_arg(*ap, int), *struct_, 0);
 	if (struct_->type == 'u')
-		write_type_u(va_arg(*ap, int), *struct_);
+		write_type_u(va_arg(*ap, unsigned long long int), *struct_);
 }
